# Improve the performance of the method toKubeContainerImageSpec