--- a/pkg/kubelet/kuberuntime/convert.go
+++ b/pkg/kubelet/kuberuntime/convert.go
@@ -17,8 +17,6 @@ limitations under the License.
 package kuberuntime
 
 import (
-	"sort"
-
 	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1"
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
 )
@@ -29,16 +27,11 @@ func toKubeContainerImageSpec(image *runtimeapi.Image) kubecontainer.ImageSpec {
 	var annotations []kubecontainer.Annotation
 
 	if image.Spec != nil && len(image.Spec.Annotations) > 0 {
-		annotationKeys := make([]string, 0, len(image.Spec.Annotations))
 		annotations = make([]kubecontainer.Annotation, 0, len(image.Spec.Annotations))
-		for k := range image.Spec.Annotations {
-			annotationKeys = append(annotationKeys, k)
-		}
-		sort.Strings(annotationKeys)
-		for _, k := range annotationKeys {
+		for k, v := range image.Spec.Annotations {
 			annotations = append(annotations, kubecontainer.Annotation{
 				Name:  k,
-				Value: image.Spec.Annotations[k],
+				Value: v,
 			})
 		}
 	}