@pacoxu 2023-10-08
@sourcelliu do you have some special cases with a large number of images on a node to convince that it will end up being a big problem for the node?

> My feeling is that the reproducible/predictable output is more important than a small perf gain here. I'd nack/close this one.
> 
> ping @SergeyKanzhelev @bobbypage

A quick look is that this is a kubelet internal function that is only used by image GC manager. My vote would be +1 for this if this can help to do a quicker GC for images. 

- BTW,(probably off topic) https://github.com/kubernetes/enhancements/issues/4210 is in progress to make GC better. @sourcelliu  not sure if the KEP will help in your case.